add a TimeList::list_from_range factory function for creating a TimeList from a TimeRange
The resulting TimeList will contain all times in the given range that result
from iterating the range with a step value of 1.0 in the rate of the range's
duration. As a result, the returned TimeList will share the same rate as the
given range's duration.

Note that durations with values less than or equal to zero will result in a
TimeList containing just the start time of the range, rescaled to the duration's
rate.